requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-jose
passlib[bcrypt]
python-dotenv
alembic